Add SG_Polymorph reveal to TRUESIGHT
The effect is the one from ELEMENTAL_WEAPON_THUNDER. I picked it because it pulses like the glowing eyes from TRUESIGHT, but they don't pulse at the same rate. Ideally it would be tweaked to do that and probably be a little more purplish.

Also, I was going to have TRUEFORM_REVEAL put another status visible on the portrait indicator on the target that named what their original type is, but I haven't figured out exactly how yet. Might require SE to do well.